+++ CHANGELOG.md
@@ -6,11 +6,15 @@ directly impact users rather than highlighting other crucial architectural updat
 
 ## [Unreleased]
 
+### Added
+- Insufficient funds sheet in swaps.
+
 ### Changed
 - Sending flow states have been simplified. We no longer show failure, resubmission, or partial statuses. Everything has been consolidated and is now represented by a single Pending state.
 
 ### Fixed
 - ZEC on other chains is no longer filtered out in swap flows.
+- Insufficient funds sheet layout.
 
 ## 2.4.9 build 1 (2025-12-04)